+++ src/config/mod.rs
@@ -1,8 +1,13 @@
-use log::debug;
-use log::info;
+use crate::schema::utils::{EmbeddedSchemaResolver, CONFIG_SCHEMA_STRING};
+use crate::storage::jenv::{get_env_var, set_env_var, set_env_var_override, EnvError};
+use jsonschema::{Draft, Registry, Retrieve, Validator};
+
+use log::{debug, error, info};
 use serde::Deserialize;
 use serde::Serialize;
-use std::env;
+use serde_json::Value;
+use std::error::Error;
+use std::fmt;
 use std::fs;
 use std::path::PathBuf;
 use uuid::Uuid;
@@ -10,6 +15,7 @@ use uuid::Uuid;
 #[derive(Serialize, Deserialize, Default, Debug)]
 pub struct Config {
     #[serde(rename = "$schema")]
+    #[serde(default = "default_schema")]
     schema: String,
     jacs_use_filesystem: Option<String>,
     jacs_use_security: Option<String>,
@@ -23,6 +29,11 @@ pub struct Config {
     jacs_signature_schema_version: Option<String>,
     jacs_private_key_password: Option<String>,
     jacs_agent_id_and_version: Option<String>,
+    jacs_default_storage: Option<String>,
+}
+
+fn default_schema() -> String {
+    "https://hai.ai/schemas/jacs.config.schema.json".to_string()
 }
 
 impl Config {
@@ -40,6 +51,7 @@ impl Config {
         jacs_signature_schema_version: Option<String>,
         jacs_private_key_password: Option<String>,
         jacs_agent_id_and_version: Option<String>,
+        jacs_default_storage: Option<String>,
     ) -> Config {
         Config {
             schema,
@@ -55,17 +67,63 @@ impl Config {
             jacs_signature_schema_version,
             jacs_private_key_password,
             jacs_agent_id_and_version,
+            jacs_default_storage,
         }
     }
 }
 
+impl fmt::Display for Config {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            r#"
+        Loading JACS and Sophon env variables of:
+            JACS_USE_SECURITY                {},
+            JACS_USE_FILESYSTEM:             {},
+            JACS_DATA_DIRECTORY:             {},
+            JACS_KEY_DIRECTORY:              {},
+            JACS_AGENT_PRIVATE_KEY_FILENAME: {},
+            JACS_AGENT_PUBLIC_KEY_FILENAME:  {},
+            JACS_AGENT_KEY_ALGORITHM:        {},
+            JACS_AGENT_SCHEMA_VERSION:       {},
+            JACS_HEADER_SCHEMA_VERSION:      {},
+            JACS_SIGNATURE_SCHEMA_VERSION:   {},
+            JACS_PRIVATE_KEY_PASSWORD        {},
+            JACS_AGENT_ID_AND_VERSION        {}
+        "#,
+            self.jacs_use_security.as_deref().unwrap_or(""),
+            self.jacs_use_filesystem.as_deref().unwrap_or(""),
+            self.jacs_data_directory.as_deref().unwrap_or(""),
+            self.jacs_key_directory.as_deref().unwrap_or(""),
+            self.jacs_agent_private_key_filename
+                .as_deref()
+                .unwrap_or(""),
+            self.jacs_agent_public_key_filename.as_deref().unwrap_or(""),
+            self.jacs_agent_key_algorithm.as_deref().unwrap_or(""),
+            self.jacs_agent_schema_version.as_deref().unwrap_or(""),
+            self.jacs_header_schema_version.as_deref().unwrap_or(""),
+            self.jacs_signature_schema_version.as_deref().unwrap_or(""),
+            self.jacs_private_key_password.as_deref().unwrap_or(""),
+            self.jacs_agent_id_and_version.as_deref().unwrap_or(""),
+        )
+    }
+}
+
 pub fn get_default_dir() -> PathBuf {
-    env::var("JACS_DATA_DIRECTORY")
-        .map(PathBuf::from)
-        .unwrap_or_else(|_| {
-            env::set_var("JACS_DATA_DIRECTORY", ".");
-            env::current_dir().unwrap_or_else(|_| PathBuf::from("."))
-        })
+    match get_env_var("JACS_DATA_DIRECTORY", false) {
+        Ok(Some(dir)) => PathBuf::from(dir),
+        _ => {
+            #[cfg(not(target_arch = "wasm32"))]
+            {
+                let _ = set_env_var("JACS_DATA_DIRECTORY", ".");
+                std::env::current_dir().unwrap_or_else(|_| PathBuf::from("."))
+            }
+            #[cfg(target_arch = "wasm32")]
+            {
+                PathBuf::from(".")
+            }
+        }
+    }
 }
 
 pub fn split_id(input: &str) -> Option<(&str, &str)> {
@@ -82,92 +140,166 @@ pub fn split_id(input: &str) -> Option<(&str, &str)> {
     }
 }
 
-pub fn set_env_vars() -> String {
-    let config: Config = match fs::read_to_string("jacs.config.json") {
-        Ok(content) => serde_json::from_str(&content).unwrap_or_default(),
-        Err(_) => Config {
-            schema: "https://hai.ai/schemas/jacs.config.schema.json".to_string(),
-            jacs_use_filesystem: None,
-            jacs_use_security: None,
-            jacs_data_directory: None,
-            jacs_key_directory: None,
-            jacs_agent_private_key_filename: None,
-            jacs_agent_public_key_filename: None,
-            jacs_agent_key_algorithm: None,
-            jacs_agent_schema_version: None,
-            jacs_header_schema_version: None,
-            jacs_signature_schema_version: None,
-            jacs_private_key_password: None,
-            jacs_agent_id_and_version: None,
+pub fn validate_config(config_json: &str) -> Result<Value, Box<dyn Error>> {
+    let jacsconfigschema_result: Value = serde_json::from_str(CONFIG_SCHEMA_STRING)?;
+
+    let jacsconfigschema = Validator::options()
+        .with_draft(Draft::Draft7)
+        .with_retriever(EmbeddedSchemaResolver::new())
+        .build(&jacsconfigschema_result)?;
+
+    let instance: Value = serde_json::from_str(config_json).map_err(|e| {
+        error!("Invalid JSON: {}", e);
+        e
+    })?;
+
+    debug!("validate json {:?}", instance);
+
+    // Validate and map any error into an owned error (a boxed String error).
+    jacsconfigschema.validate(&instance).map_err(|e| {
+        let err_msg = format!("Error validating config file: {}", e);
+        error!("{}", err_msg);
+        Box::<dyn Error>::from(err_msg)
+    })?;
+
+    Ok(instance)
+}
+
+// THis function should be exposed to higher level functions that can load configs from
+// anywere. Additionally, since this is called automatically in creating an agent
+// it should not override any existing env vars.
+pub fn set_env_vars(
+    do_override: bool,
+    config_json: Option<&str>,
+) -> Result<String, Box<dyn Error>> {
+    let config: Config = match config_json {
+        Some(json_str) => serde_json::from_value(validate_config(json_str).unwrap_or_default())
+            .unwrap_or_default(),
+        None => match fs::read_to_string("jacs.config.json") {
+            Ok(content) => serde_json::from_value(validate_config(&content).unwrap_or_default())
+                .unwrap_or_default(),
+            Err(_) => Config::default(),
         },
     };
     debug!("configs from file {:?}", config);
-
+    validate_config(&serde_json::to_string(&config).map_err(|e| Box::new(e) as Box<dyn Error>)?)?;
     let jacs_use_filesystem = config
         .jacs_use_filesystem
-        .unwrap_or_else(|| "true".to_string());
-    env::set_var("JACS_USE_FILESYSTEM", &jacs_use_filesystem);
+        .as_ref()
+        .unwrap_or(&"true".to_string())
+        .clone();
+    set_env_var_override("JACS_USE_FILESYSTEM", &jacs_use_filesystem, do_override)?;
 
     let jacs_private_key_password = config
         .jacs_private_key_password
-        .unwrap_or_else(|| "true".to_string());
-    env::set_var("JACS_PRIVATE_KEY_PASSWORD", &jacs_private_key_password);
+        .as_ref()
+        .unwrap_or(&"true".to_string())
+        .clone();
+    set_env_var_override(
+        "JACS_PRIVATE_KEY_PASSWORD",
+        &jacs_private_key_password,
+        do_override,
+    )?;
 
     let jacs_use_security = config
         .jacs_use_security
-        .unwrap_or_else(|| "false".to_string());
-    env::set_var("JACS_USE_SECURITY", &jacs_use_security);
+        .as_ref()
+        .unwrap_or(&"false".to_string())
+        .clone();
+    set_env_var_override("JACS_USE_SECURITY", &jacs_use_security, do_override)?;
 
     let jacs_data_directory = config
         .jacs_data_directory
-        .unwrap_or_else(|| format!("{:?}", env::current_dir().unwrap()));
-    env::set_var("JACS_DATA_DIRECTORY", &jacs_data_directory);
+        .as_ref()
+        .unwrap_or(&format!("{:?}", std::env::current_dir().unwrap()))
+        .clone();
+    set_env_var_override("JACS_DATA_DIRECTORY", &jacs_data_directory, do_override)?;
 
-    let jacs_key_directory = config.jacs_key_directory.unwrap_or_else(|| ".".to_string());
-    env::set_var("JACS_KEY_DIRECTORY", &jacs_key_directory);
+    let jacs_key_directory = config
+        .jacs_key_directory
+        .as_ref()
+        .unwrap_or(&".".to_string())
+        .clone();
+    set_env_var_override("JACS_KEY_DIRECTORY", &jacs_key_directory, do_override)?;
 
     let jacs_agent_private_key_filename = config
         .jacs_agent_private_key_filename
-        .unwrap_or_else(|| "rsa_pss_private.pem".to_string());
-    env::set_var(
+        .as_ref()
+        .unwrap_or(&"rsa_pss_private.pem".to_string())
+        .clone();
+    set_env_var_override(
         "JACS_AGENT_PRIVATE_KEY_FILENAME",
         &jacs_agent_private_key_filename,
-    );
+        do_override,
+    )?;
 
     let jacs_agent_public_key_filename = config
         .jacs_agent_public_key_filename
-        .unwrap_or_else(|| "rsa_pss_public.pem".to_string());
-    env::set_var(
+        .as_ref()
+        .unwrap_or(&"rsa_pss_public.pem".to_string())
+        .clone();
+    set_env_var_override(
         "JACS_AGENT_PUBLIC_KEY_FILENAME",
         &jacs_agent_public_key_filename,
-    );
+        do_override,
+    )?;
 
     let jacs_agent_key_algorithm = config
         .jacs_agent_key_algorithm
-        .unwrap_or_else(|| "RSA-PSS".to_string());
-    env::set_var("JACS_AGENT_KEY_ALGORITHM", &jacs_agent_key_algorithm);
+        .as_ref()
+        .unwrap_or(&"RSA-PSS".to_string())
+        .clone();
+    set_env_var_override(
+        "JACS_AGENT_KEY_ALGORITHM",
+        &jacs_agent_key_algorithm,
+        do_override,
+    )?;
 
     let jacs_agent_schema_version = config
         .jacs_agent_schema_version
-        .unwrap_or_else(|| "v1".to_string());
-    env::set_var("JACS_AGENT_SCHEMA_VERSION", &jacs_agent_schema_version);
+        .as_ref()
+        .unwrap_or(&"v1".to_string())
+        .clone();
+    set_env_var_override(
+        "JACS_AGENT_SCHEMA_VERSION",
+        &jacs_agent_schema_version,
+        do_override,
+    )?;
 
     let jacs_header_schema_version = config
         .jacs_header_schema_version
-        .unwrap_or_else(|| "v1".to_string());
-    env::set_var("JACS_HEADER_SCHEMA_VERSION", &jacs_header_schema_version);
+        .as_ref()
+        .unwrap_or(&"v1".to_string())
+        .clone();
+    set_env_var_override(
+        "JACS_HEADER_SCHEMA_VERSION",
+        &jacs_header_schema_version,
+        do_override,
+    )?;
 
     let jacs_signature_schema_version = config
         .jacs_signature_schema_version
-        .unwrap_or_else(|| "v1".to_string());
-    env::set_var(
+        .as_ref()
+        .unwrap_or(&"v1".to_string())
+        .clone();
+    set_env_var_override(
         "JACS_SIGNATURE_SCHEMA_VERSION",
         &jacs_signature_schema_version,
-    );
+        do_override,
+    )?;
+
+    let jacs_default_storage = config
+        .jacs_default_storage
+        .as_ref()
+        .unwrap_or(&"fs".to_string())
+        .clone();
+    set_env_var_override("JACS_DEFAULT_STORAGE", &jacs_default_storage, do_override)?;
 
     let jacs_agent_id_and_version = config
         .jacs_agent_id_and_version
-        .unwrap_or_else(|| "".to_string());
+        .as_ref()
+        .unwrap_or(&"".to_string())
+        .clone();
 
     if !jacs_agent_id_and_version.is_empty() {
         let (id, version) = split_id(&jacs_agent_id_and_version).unwrap_or(("", ""));
@@ -176,38 +308,64 @@ pub fn set_env_vars() -> String {
         }
     }
 
-    env::set_var("JACS_AGENT_ID_AND_VERSION", &jacs_agent_id_and_version);
+    set_env_var_override(
+        "JACS_AGENT_ID_AND_VERSION",
+        &jacs_agent_id_and_version,
+        do_override,
+    )?;
 
-    let loading_message = format!(
-        r#"
-        Loading JACS and Sophon env variables of:
-            JACS_USE_SECURITY                {},
-            JACS_USE_FILESYSTEM:             {},
-            JACS_DATA_DIRECTORY:             {},
-            JACS_KEY_DIRECTORY:              {},
-            JACS_AGENT_PRIVATE_KEY_FILENAME: {},
-            JACS_AGENT_PUBLIC_KEY_FILENAME:  {},
-            JACS_AGENT_KEY_ALGORITHM:        {},
-            JACS_AGENT_SCHEMA_VERSION:       {},
-            JACS_HEADER_SCHEMA_VERSION:      {},
-            JACS_SIGNATURE_SCHEMA_VERSION:   {},
-            JACS_PRIVATE_KEY_PASSWORD        {},
-            JACS_AGENT_ID_AND_VERSION        {}
-        "#,
-        jacs_use_security,
-        jacs_use_filesystem,
-        jacs_data_directory,
-        jacs_key_directory,
-        jacs_agent_private_key_filename,
-        jacs_agent_public_key_filename,
-        jacs_agent_key_algorithm,
-        jacs_agent_schema_version,
-        jacs_header_schema_version,
-        jacs_signature_schema_version,
-        jacs_private_key_password,
-        jacs_agent_id_and_version,
-    );
+    let message = format!("{}", config);
+    info!("{}", message);
+    check_env_vars().map_err(|e| {
+        error!("Error checking environment variables: {}", e);
+        Box::new(e) as Box<dyn Error>
+    })?;
+    Ok(message)
+}
 
-    info!("{}", loading_message);
-    loading_message
+pub fn check_env_vars() -> Result<String, EnvError> {
+    let vars = [
+        ("JACS_USE_SECURITY", true),
+        ("JACS_USE_FILESYSTEM", true),
+        ("JACS_DATA_DIRECTORY", true),
+        ("JACS_KEY_DIRECTORY", true),
+        ("JACS_AGENT_PRIVATE_KEY_FILENAME", true),
+        ("JACS_AGENT_PUBLIC_KEY_FILENAME", true),
+        ("JACS_AGENT_KEY_ALGORITHM", true),
+        ("JACS_AGENT_SCHEMA_VERSION", true),
+        ("JACS_HEADER_SCHEMA_VERSION", true),
+        ("JACS_SIGNATURE_SCHEMA_VERSION", true),
+        ("JACS_PRIVATE_KEY_PASSWORD", true),
+        ("JACS_AGENT_ID_AND_VERSION", true),
+    ];
+
+    let mut message = String::from("\nChecking JACS environment variables:\n");
+    let mut missing_vars = Vec::new();
+
+    for (var_name, required) in vars.iter() {
+        let value = get_env_var(var_name, *required)?;
+        let status = match value {
+            Some(val) => val,
+            None => {
+                if *required {
+                    missing_vars.push(var_name);
+                }
+                "MISSING".to_string()
+            }
+        };
+        message.push_str(&format!(
+            "    {:<35} {}\n",
+            var_name.to_string() + ":",
+            status
+        ));
+    }
+
+    if !missing_vars.is_empty() {
+        message.push_str("\nMissing required environment variables:\n");
+        for var in missing_vars {
+            message.push_str(&format!("    {}\n", var));
+        }
+    }
+
+    Ok(message)
 }